Remove disable SetupIgnoreSSLCheck() in example project
It doesn't work anyway (did it ever?!). The way to make this work is to
support passing in a custom HttpClient to the Comparisons instance. This
HttpClient could be configured with custom certificate and hostname
validation behaviour. Note that even if an untrusted certificate is
added to the Java Key Store, that still won't pass hostname validation,
which is performed separately to certificate chain validation.